Fix some func

+++ app/Http/Controllers/Admin/ColorController.php
@@ -6,6 +6,7 @@ use Throwable;
 use App\Models\Item;
 use App\Models\Color;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\ColorResource;
@@ -13,55 +14,70 @@ use App\Http\Requests\admin\ColorRequest;
 
 class ColorController extends Controller
 {
-    // 該当のカラム以外を扱わないようにホワイトリスト作成
-    private $form_items = [ 'id', 'color_name' ];
+    private $form_items = ['id', 'color_name'];
 
     public function __construct()
     {
-        // Auth認証
         $this->middleware('auth:admin');
     }
 
     public function index(Request $request)
     {
-        // レスポンスを返却
-        return response()->json(['colors' => ColorResource::collection(Color::all())],200);
+        try {
+            return response()->json(['colors' => ColorResource::collection(Color::all())]);
+        } catch (Throwable $e) {
+            Log::error($e->getMessage());
+            return response()->json(['status' => 9, 'message' => 'カラーマスタの取得に失敗しました'], 500);
+        }
     }
 
     public function store(ColorRequest $request)
     {
-        // 不正な入力値の制御
-        $data = $request->only($this->form_items);
-        // DBに登録
-        Color::create([
-            'color_name' => $data['color_name'],
-        ]);
-        // レスポンスを返却
-        return response()->json(['create' => true, 'message' => 'カラーマスタの新規登録を完了しました'], 200);
+        DB::beginTransaction();
+        try {
+            $data = $request->only($this->form_items);
+            Color::create([
+                'color_name' => $data['color_name'],
+            ]);
+            DB::commit();
+            return response()->json(['status' => 1, 'message' => 'カラーマスタの登録を完了しました'], 200);
+        } catch (Throwable $e) {
+            Log::error($e->getMessage());
+            DB::rollBack();
+            return response()->json(['status' => 9, 'message' => 'カラーマスタの登録に失敗しました'], 500);
+        }
     }
 
     public function update(ColorRequest $request, Color $color)
     {
-        // 項目制限
-        $data = $request->only($this->form_items);
-        // 編集項目をDBに保存
-        $color->fill($data)->save();
-        // レスポンスを返却
-        return response()->json(['update' => true, 'message' => 'カラーマスタの編集を完了しました'], 200);
+        DB::beginTransaction();
+        try {
+            $data = $request->only($this->form_items);
+            $color->fill($data)->save();
+            DB::commit();
+            return response()->json(['status' => 1, 'message' => 'カラーマスタの編集を完了しました'], 200);
+        } catch (Throwable $e) {
+            Log::error($e->getMessage());
+            DB::rollBack();
+            return response()->json(['status' => 9, 'message' => 'カラーマスタの編集に失敗しました'], 500);
+        }
     }
 
     public function destroy(Color $color)
     {
+        DB::beginTransaction();
         try {
+            $related_items_arr = Item::whereIn('id', $color->skus->pluck('item_id'))->pluck('product_number')->toArray();
+            if (!empty($related_items_arr)) {
+                return response()->json(['status' => 9, 'message' => '選択カラーが商品で使用されております'], 400);
+            }
             $color->delete();
-            // レスポンスを返却
-            return response()->json(['delete' => true, 'message' => 'カラーマスタの削除を完了しました'], 200);
+            DB::commit();
+            return response()->json(['status' => 1, 'message' => 'カラーマスタの削除を完了しました'], 200);
         } catch (Throwable $e) {
             Log::error($e->getMessage());
-            $related_items_id = $color->skus->pluck('item_id');
-            $product_numbers = implode(' / ', Item::whereIn('id', $related_items_id)->pluck('product_number')->toArray());
-            // レスポンスを返却
-            return response()->json(['delete' => false, 'message' => 'カラーの削除を失敗しました。以下の品番の商品で選択カラーが使用されております。 商品マスタを編集画面にて選択カラーを変更してください。'.$product_numbers], 405);
+            DB::rollBack();
+            return response()->json(['status' => 9, 'message' => 'カラーマスタの削除を失敗しました'], 500);
         }
     }
 }
